requirements.txt
tqdm
pypub3
selenium
requests
beautifulsoup4
inquirer
coverage
EbookLib